+++ README.md
@@ -5,3 +5,21 @@
 
 ## 과제 제출 과정
 * [과제 제출 방법](https://github.com/next-step/nextstep-docs/tree/master/precourse)
+
+##기능 구현
+* 1~9 사이의 숫자 중 랜덤으로 3개의 숫자를 구한다
+  * Random method 사용
+  * 1 ~ 9 사이의 세자리 숫자
+  * 각 자리 숫자는 중복 되지 않는다.
+* 사용자로 부터 입력 받는 3개 숫자 예외 처리
+  * 1 ~ 9 사이의 숫자인가?
+  * 중복 값이 있는가?
+  * 3자리 숫자 인가?
+  * 세자루가 아닐 경우 throw IllegalArgumentException
+* 위치와 숫자가 값이 같은 경우 - 스트라이크
+* 위치는 다른데 숫자 값이 같은 경우 - 볼
+* 숫자 값이 없는 경우 - 낫싱
+* 사용자가 입력한 값에 대한 실행 결과를 구한다
+  * 안내 메세지 출력
+  * 1을 입력할 경우 새게임 시작
+    * 1이외의 숫자는 프로그램 종료
