+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.play.GameStart;
+import baseball.start.GameStart;
 
 public class Application {
     public static void main(String[] args) {
